PM-31 reused pet list for shelter page
