+++ src/routes/SheltersPege/SheltersPege.js
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
-import { Section, FormGroup } from '../../components/Utils/Utils';
+import { Section } from '../../components/Utils/Utils';
 import './SheltersPege.css';
 import SheltersService from '../../services/sheltersService';
-import PetCard from '../../components/PetCard';
 import * as logUtils from '../../components/Utils/Logger';
+import PetList from '../../components/PetList';
+import PetsService from '../../services/petsService';
 var log = logUtils.getLogger()
 
 export default class SheltersPege extends Component {
@@ -14,24 +15,65 @@ export default class SheltersPege extends Component {
       shelterID: null,
       shelter: null,
       pets: null,
-      typeOfPets: "Any",
-      age: "Any",
-      availability: "Any"
+      typeOfAnimal: [],
+      breed: [],
+      sex: '',
+      minAge: '',
+      maxAge: '',
+      more: []
     };
+    this.inputChanged = this.inputChanged.bind(this);
   }
 
   componentDidMount() {
     log.debug("Getting shelter data...")
     const path = window.location.pathname;
     const shelterID = path.substring(path.lastIndexOf('/') + 1)
+    this.setState({ shelterID });
     SheltersService.getShelter(shelterID)
       .then(shelter => this.setState({ shelter: shelter }))
       .catch(error => console.log(error));
 
     log.debug("Getting pets data...")
-    SheltersService.getPetsByShelter(shelterID)
+    PetsService.getPets({ shelterID })
       .then(pets => this.setState({ pets: pets }))
       .catch(error => console.log(error));
+  }
+
+  inputChanged(field, content) {
+    log.debug("Calling inputChanged")
+    this.setState({ [field]: content }, () => this.handleSubmit());
+  }
+
+  handleSubmit = () => {
+    PetsService.getPets(this.getFilters())
+      .then(pets => this.setState({ pets: pets }))
+      .catch(error => console.log(error));
+}
+
+  getFilters() {
+    const {
+      shelterID,
+      typeOfAnimal,
+      breed,
+      sex,
+      minAge,
+      maxAge,
+      more
+    } = this.state;
+    const filters = { shelterID, availability: 'Available' };
+    if (typeOfAnimal.length > 0)
+      filters.typeOfAnimal = Array.from(typeOfAnimal, option => option.value);
+    if (breed.length > 0)
+      filters.breed = Array.from(breed, option => option.value);
+    if (sex)
+      filters.sex = sex.value;
+    if (minAge > 0)
+      filters.minAge = minAge;
+    if (maxAge > 0)
+      filters.maxAge = maxAge;
+    more.forEach(option => filters[option.value] = true);
+    return filters;
   }
 
   renderShelter(shelter) {
@@ -68,117 +110,44 @@ export default class SheltersPege extends Component {
     );
   }
 
-  renderPet(pet, cName) {
-    log.debug("Calling renderPet")
+  renderPetList() {
+    log.debug("Calling renderPetList")
+    const {
+      pets,
+      typeOfAnimal,
+      breed,
+      sex,
+      minAge,
+      maxAge,
+      more
+    } = this.state;
     return (
-      <PetCard className={`petCard ${cName}`} key={pet.PetID} pet={pet} page='shelter' />
+      <>
+        {pets && <PetList
+          pets={pets}
+          page='shelter'
+          typeOfAnimal={typeOfAnimal}
+          breed={breed}
+          sex={sex}
+          minAge={minAge}
+          maxAge={maxAge}
+          more={more}
+          inputChangeHandler={this.inputChanged}
+        />}
+      </>
     );
-  }
-
-  renderPets(pets, typeOfPets, age, available) {
-    log.debug("Calling renderPets")
-    // no pets exist
-    if (pets.length === 0) {
-      return (
-        <div className='alert alert-success' role='alert'>
-          More pets will be available soon! Check back later :)
-        </div>
-      );
-    } else {
-      return (
-        <div>
-          <Section>
-            {this.renderFilter()}
-          </Section>
-          <Section>
-            <div className='row'>
-              {pets.map(pet => {
-                let typeOfAnimal = (pet.TypeOfAnimal === "Cat" || pet.TypeOfAnimal === "Dog") ? pet.TypeOfAnimal : "Other"
-
-                let petAge = ""
-                if (parseInt(pet.Age) <= 1 && typeOfAnimal === "Cat") petAge = "Kitten"
-                else if (parseInt(pet.Age) <= 1 && typeOfAnimal === "Dog") petAge = "Puppy"
-                else if (parseInt(pet.Age) <= 2) petAge = "Young"
-                else if (parseInt(pet.Age) <= 7) petAge = "Adult"
-                else petAge = "Senior"
-
-                let cName = "hideCard"
-                if (typeOfPets === "Any" && age === "Any" && available === "Any") cName = "showCard"
-                if (typeOfPets === "Any" && age === "Any" && available === pet.Availability) cName = "showCard"
-                if (typeOfPets === "Any" && age === petAge && available === pet.Availability) cName = "showCard"
-                if (typeOfPets === typeOfAnimal && age === "Any" && available === "Any") cName = "showCard"
-                if (typeOfPets === typeOfAnimal && age === petAge && available === "Any") cName = "showCard"
-                if (typeOfPets === "Any" && age === petAge && available === "Any") cName = "showCard"
-                if (typeOfPets === "Any" && age === petAge && available === pet.Availability) cName = "showCard"
-                if (typeOfPets === typeOfAnimal && age === "Any" && available === pet.Availability) cName = "showCard"
-                if (typeOfPets === typeOfAnimal && age === petAge && available === pet.Availability) cName = "showCard"
-
-                return this.renderPet(pet, cName)
-              })}
-            </div>
-          </Section>
-
-        </div>
-
-      )
-    }
-  }
-
-  renderFilter() {
-    log.debug("Calling renderFilter")
-    return (
-      <FormGroup className="flex">
-        <div className="select-wrap" style={{ width: "100%" }}>
-          <label>Type of Pets</label>
-          <select id='typeOfPets' name="typeOfPets" onChange={e => this.inputChanged('typeOfPets', e.target.value)}>
-            <option value='Any'>Any</option>
-            <option value='Cat'>Cat</option>
-            <option value='Dog'>Dog</option>
-            <option value='Other'>Other</option>
-          </select>
-        </div>
-
-        <div className="select-wrap" style={{ width: "100%" }}>
-          <label>Age</label>
-          <select id='age' name="age" onChange={e => this.inputChanged('age', e.target.value)}>
-            <option value='Any'>Any</option>
-            <option value='Puppy'>Puppy</option>
-            <option value='Kitten'>Kitten</option>
-            <option value='Young'>Young</option>
-            <option value='Adult'>Adult</option>
-            <option value='Senior'>Senior</option>
-          </select>
-        </div>
-
-        <div className="select-wrap" style={{ width: "100%" }}>
-          <label>Availability</label>
-          <select id='availability' name="availability" onChange={e => this.inputChanged('availability', e.target.value)} >
-            <option value='Any'>Any</option>
-            <option value='Available'>Available</option>
-            <option value='Pending'>Pending</option>
-          </select>
-        </div>
-
-      </FormGroup>
-    )
-  }
-
-  inputChanged(field, content) {
-    log.debug("Calling inputChanged")
-    this.setState({ [field]: content });
   }
 
   render() {
     log.debug("Calling render...")
-    const { shelter, pets } = this.state;
-
+    const { shelter } = this.state;
     return (
       <div>
         <Section>
-          {this.state.shelter == null ? null : this.renderShelter(shelter)}
+          {shelter == null ? null : this.renderShelter(shelter)}
         </Section>
         <Section>
-          {this.state.pets == null ? null : this.renderPets(pets, this.state.typeOfPets, this.state.age, this.state.availability)}
+          {this.renderPetList()}
         </Section>
       </div>
